requirements.txt
faster-whisper==1.0.3
opencv-python==4.10.0.84
Pillow==10.4.0
pytesseract==0.3.10
jieba==0.42.1
srt==3.5.3
numpy==1.26.4
